Revert "update spaceflights tutorial"
