+++ spaceflights-kedro/.gitignore
@@ -4,7 +4,6 @@
 # ignore all local configuration
 conf/local/**
 !conf/local/.gitkeep
-.telemetry
 
 # ignore potentially sensitive credentials files
 conf/**/*credentials*
@@ -20,8 +19,8 @@ logs/**
 # also keep all .gitkeep files
 !.gitkeep
 
-# also keep the example dataset
-!data/01_raw/iris.csv
+# keep also the example dataset
+!data/01_raw/*
 
 
 ##########################
