+++ spaceflights-kedro/.ipython/profile_default/startup/00-kedro-init.py
@@ -4,27 +4,11 @@ from pathlib import Path
 from typing import Any, Dict
 
 from IPython.core.magic import needs_local_scope, register_line_magic
-
 # Find the project root (./../../../)
+from kedro.framework.startup import _get_project_metadata
+
 startup_error = None
 project_path = Path(__file__).parents[3].resolve()
-
-
-def _remove_cached_modules(package_name):
-    to_remove = [mod for mod in sys.modules if mod.startswith(package_name)]
-    # `del` is used instead of `reload()` because: If the new version of a module does not
-    # define a name that was defined by the old version, the old definition remains.
-    for module in to_remove:
-        del sys.modules[module]  # pragma: no cover
-
-
-def _clear_hook_manager():
-    from kedro.framework.hooks import get_hook_manager
-
-    hook_manager = get_hook_manager()
-    name_plugin_pairs = hook_manager.list_name_plugin()
-    for name, plugin in name_plugin_pairs:
-        hook_manager.unregister(name=name, plugin=plugin)  # pragma: no cover
 
 
 @register_line_magic
@@ -38,9 +22,10 @@ def reload_kedro(path, line=None, env: str = None, extra_params: Dict[str, Any] 
     try:
         import kedro.config.default_logger
         from kedro.framework.cli.jupyter import collect_line_magic
+        from kedro.framework.hooks import get_hook_manager
+        from kedro.framework.project import configure_project
         from kedro.framework.session import KedroSession
         from kedro.framework.session.session import _activate_session
-        from kedro.framework.startup import bootstrap_project
     except ImportError:
         logging.error(
             "Kedro appears not to be installed in your current environment "
@@ -49,13 +34,25 @@ def reload_kedro(path, line=None, env: str = None, extra_params: Dict[str, Any] 
         raise
 
     try:
-        _clear_hook_manager()
-
         path = path or project_path
-        metadata = bootstrap_project(path)
 
-        _remove_cached_modules(metadata.package_name)
+        # clear hook manager
+        hook_manager = get_hook_manager()
+        name_plugin_pairs = hook_manager.list_name_plugin()
+        for name, plugin in name_plugin_pairs:
+            hook_manager.unregister(name=name, plugin=plugin)
 
+        # remove cached user modules
+        metadata = _get_project_metadata(path)
+        to_remove = [
+            mod for mod in sys.modules if mod.startswith(metadata.package_name)
+        ]
+        # `del` is used instead of `reload()` because: If the new version of a module does not
+        # define a name that was defined by the old version, the old definition remains.
+        for module in to_remove:
+            del sys.modules[module]
+
+        configure_project(metadata.package_name)
         session = KedroSession.create(
             metadata.package_name, path, env=env, extra_params=extra_params
         )
