+++ spaceflights-kedro/conf/base/catalog.yml
@@ -2,45 +2,77 @@
 #
 # Documentation for this file format can be found in "The Data Catalog"
 # Link: https://kedro.readthedocs.io/en/stable/05_data/01_data_catalog.html
+#
+# We support interacting with a variety of data stores including local file systems, cloud, network and HDFS
+#
+# An example data set definition can look as follows:
+#
+#bikes:
+#  type: pandas.CSVDataSet
+#  filepath: "data/01_raw/bikes.csv"
+#
+#weather:
+#  type: spark.SparkDataSet
+#  filepath: s3a://your_bucket/data/01_raw/weather*
+#  file_format: csv
+#  credentials: dev_s3
+#  load_args:
+#    header: True
+#    inferSchema: True
+#  save_args:
+#    sep: '|'
+#    header: True
+#
+#scooters:
+#  type: pandas.SQLTableDataSet
+#  credentials: scooters_credentials
+#  table_name: scooters
+#  load_args:
+#    index_col: ['name']
+#    columns: ['name', 'gear']
+#  save_args:
+#    if_exists: 'replace'
+#    # if_exists: 'fail'
+#    # if_exists: 'append'
+#
+# The Data Catalog supports being able to reference the same file using two different DataSet implementations
+# (transcoding), templating and a way to reuse arguments that are frequently repeated. See more here:
+# https://kedro.readthedocs.io/en/stable/05_data/01_data_catalog.html
 
 companies:
   type: pandas.CSVDataSet
-  filepath: data/01_raw/Allcompanies.csv
+  filepath: data/01_raw/companies.csv
+  # more about layers in the Data Engineering Convention:
+  # https://kedro.readthedocs.io/en/stable/03_tutorial/06_visualise_pipeline.html#interact-with-data-engineering-convention
   layer: raw
 
 reviews:
   type: pandas.CSVDataSet
   filepath: data/01_raw/reviews.csv
   layer: raw
-  
+
 shuttles:
   type: pandas.ExcelDataSet
   filepath: data/01_raw/shuttles.xlsx
   layer: raw
-  
+
 preprocessed_companies:
   type: pandas.CSVDataSet
   filepath: data/02_intermediate/preprocessed_companies.csv
-  load_args:
-    encoding: 'latin_1'
   layer: intermediate
-  
+
 preprocessed_shuttles:
   type: pandas.CSVDataSet
   filepath: data/02_intermediate/preprocessed_shuttles.csv
   layer: intermediate
-    
+
 model_input_table:
   type: pandas.CSVDataSet
   filepath: data/03_primary/model_input_table.csv
-  load_args:
-    encoding: 'latin_1'
-  layer: Primary
+  layer: primary
 
 regressor:
   type: pickle.PickleDataSet
   filepath: data/06_models/regressor.pickle
-  load_args:
-    encoding: 'latin_1'
   versioned: true
   layer: models
