+++ spaceflights-kedro/docs/source/index.rst
@@ -1,8 +1,8 @@
-.. space2 documentation master file, created by sphinx-quickstart.
+.. spaceflights documentation master file, created by sphinx-quickstart.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to project space2's API docs!
+Welcome to project spaceflights's API docs!
 =============================================
 
 .. toctree::
