--- spaceflights-kedro/docs/source/modules.rst
@@ -1,7 +1,0 @@
-space2
-======
-
-.. toctree::
-   :maxdepth: 4
-
-   space2
