--- spaceflights-kedro/docs/source/space2.pipelines.data_processing.rst
@@ -1,26 +1,0 @@
-space2.pipelines.data\_processing package
-=========================================
-
-.. automodule:: space2.pipelines.data_processing
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Submodules
-----------
-
-space2.pipelines.data\_processing.nodes module
-----------------------------------------------
-
-.. automodule:: space2.pipelines.data_processing.nodes
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-space2.pipelines.data\_processing.pipeline module
--------------------------------------------------
-
-.. automodule:: space2.pipelines.data_processing.pipeline
-   :members:
-   :undoc-members:
-   :show-inheritance:
