--- spaceflights-kedro/docs/source/space2.pipelines.data_science.rst
@@ -1,26 +1,0 @@
-space2.pipelines.data\_science package
-======================================
-
-.. automodule:: space2.pipelines.data_science
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Submodules
-----------
-
-space2.pipelines.data\_science.nodes module
--------------------------------------------
-
-.. automodule:: space2.pipelines.data_science.nodes
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-space2.pipelines.data\_science.pipeline module
-----------------------------------------------
-
-.. automodule:: space2.pipelines.data_science.pipeline
-   :members:
-   :undoc-members:
-   :show-inheritance:
