--- spaceflights-kedro/docs/source/space2.pipelines.rst
@@ -1,16 +1,0 @@
-space2.pipelines package
-========================
-
-.. automodule:: space2.pipelines
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   space2.pipelines.data_processing
-   space2.pipelines.data_science
