--- spaceflights-kedro/docs/source/space2.rst
@@ -1,50 +1,0 @@
-space2 package
-==============
-
-.. automodule:: space2
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   space2.pipelines
-
-Submodules
-----------
-
-space2.cli module
------------------
-
-.. automodule:: space2.cli
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-space2.hooks module
--------------------
-
-.. automodule:: space2.hooks
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-space2.pipeline\_registry module
---------------------------------
-
-.. automodule:: space2.pipeline_registry
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-space2.settings module
-----------------------
-
-.. automodule:: space2.settings
-   :members:
-   :undoc-members:
-   :show-inheritance:
