--- spaceflights-kedro/src/space2/pipelines/data_processing/__init__.py
@@ -1,1 +1,0 @@
-from .pipeline import create_pipeline  # NOQA
