+++ spaceflights-kedro/src/spaceflights/pipeline_registry.py
@@ -4,7 +4,7 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-# http://www.apache.org/licenses/LICENSE-2.0
+#     http://www.apache.org/licenses/LICENSE-2.0
 #
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
 # EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
@@ -19,36 +19,34 @@
 # trademarks of QuantumBlack. The License does not grant you any right or
 # license to the QuantumBlack Trademarks. You may not use the QuantumBlack
 # Trademarks or any confusingly similar mark as a trademark for your product,
-#     or use the QuantumBlack Trademarks in any other manner that might cause
+# or use the QuantumBlack Trademarks in any other manner that might cause
 # confusion in the marketplace, including but not limited to in advertising,
 # on websites, or on software.
 #
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""
-This module contains an example test.
+"""Project pipelines."""
+from typing import Dict
 
-Tests should be placed in ``src/tests``, in modules that mirror your
-project's structure, and in files named test_*.py. They are simply functions
-named ``test_*`` which test a unit of logic.
+from kedro.pipeline import Pipeline
 
-To run the tests, run ``kedro test``.
-"""
-from pathlib import Path
-
-import pytest
-from kedro.framework.context import KedroContext
+from spaceflights.pipelines import data_processing as dp
+from spaceflights.pipelines import data_science as ds
 
 
-@pytest.fixture
-def project_context():
-    return KedroContext(package_name="spaceflights", project_path=Path.cwd())
+def register_pipelines() -> Dict[str, Pipeline]:
+    """Register the project's pipelines.
 
+    Returns:
+        A mapping from a pipeline name to a ``Pipeline`` object.
 
-# The tests below are here for the demonstration purpose
-# and should be replaced with the ones testing the project
-# functionality
-class TestProjectContext:
-    def test_package_name(self, project_context):
-        assert project_context.package_name == "spaceflights"
+    """
+    data_processing_pipeline = dp.create_pipeline()
+    data_science_pipeline = ds.create_pipeline()
+
+    return {
+        "__default__": data_processing_pipeline + data_science_pipeline,
+        "dp": data_processing_pipeline,
+        "ds": data_science_pipeline,
+    }
