+++ spaceflights-kedro/src/spaceflights/pipelines/data_science/nodes.py
@@ -25,7 +25,6 @@
 #
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
 import logging
 from typing import Dict, Tuple
